rename to better represent the intent of the code, second pass
